Positioning and Orienting the Camera
No idea why the texture needs to be flipped for this to appear
right-side up, considering all previous versions were fine with the way
it was.

I printed out all the scalars and vectors and compared them to an
earlier version of this that I did in C++, but all the numbers were
correct, so they weren't the problem.